Restructured /bin/ folder
