--- bin/routerManager.js
@@ -1,5 +1,0 @@
-export {indexRouter} from '../routes/index.js'
-export {registerRouter} from '../routes/register.js';
-export {watchRouter} from '../routes/watch.js';
-export {downloadRouter} from '../routes/download.js';
-export * from '../routes/management.js';
