+++ bin/web-server/routerManager.js
@@ -1,0 +1,5 @@
+export {indexRouter} from '../../routes/index.js'
+export {registerRouter} from '../../routes/register.js';
+export {watchRouter} from '../../routes/watch.js';
+export {downloadRouter} from '../../routes/download.js';
+export {managementRouter, deleteRouter} from '../../routes/management.js';
