+++ bin/web-server/startup-exit.js
@@ -1,12 +1,12 @@
 import fs from "fs";
-import {failedDownloads, newVideos, videoList} from "./dataFiles.js";
-import {addToQueue, tryDownload} from "./downloadManager.js";
+import {failedDownloads, newVideos, videoList} from "../fileSysem/dataFiles.js";
+import {addToQueue, tryDownload} from "../download/downloadManager.js";
 
 let cleanUpAndExit = function () {
 
     fs.writeFileSync('./data/videoData.json', JSON.stringify(videoList));
     fs.writeFileSync('./data/newVideos.json', JSON.stringify(newVideos));
-    fs.writeFileSync('./data/failedDownloads', JSON.stringify(failedDownloads));
+    fs.writeFileSync('./data/failedDownloads.json', JSON.stringify(failedDownloads));
     process.exit(0);
 };
